# api-gateway/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Employee Management System API
  description: |
    Comprehensive Employee Management System API Gateway
    
    ## Overview
    This microservices-based Employee Management System provides complete functionality for managing employees, departments, and authentication through a unified API Gateway.
    
    ## Authentication
    - **JWT Bearer Token**: All protected endpoints require a valid JWT token
    - **Test Credentials**: Use POST /api/auth/login with:
      - Username: `admin`
      - Password: `admin123`
    
    ## Services
    - **Auth Service**: User authentication and authorization
    - **Employee Service**: Employee and department management
    - **API Gateway**: Unified entry point (this documentation)
    
    ## Service URLs
    - API Gateway: http://localhost:8080
    - Auth Service: http://localhost:8082
    - Employee Service: http://localhost:8083
    
    ## Quick Start
    1. Authenticate via `/api/auth/login`
    2. Copy the JWT token from response
    3. Click "Authorize" button and enter: `Bearer YOUR_TOKEN`
    4. Explore protected endpoints
    
  version: 1.0.0
  contact:
    name: Employee Management System
    email: support@employeemgt.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: API Gateway (Primary)
  - url: http://localhost:8082
    description: Auth Service (Direct)
  - url: http://localhost:8083
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/login endpoint.
        Format: Bearer <token>
        
        **Quick Test**: Login with admin/admin123 to get a token
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        message:
          type: string
          description: Human-readable response message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response payload data
          nullable: true
      required:
        - success
        - message

    # Authentication Schemas
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: Username for authentication
          example: "admin"
        password:
          type: string
          description: Password for authentication
          example: "admin123"
          format: password
      required:
        - username
        - password
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          description: Unique username
          example: "john.doe"
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@company.com"
        password:
          type: string
          description: User password
          example: "SecurePass123!"
          format: password
          minLength: 6
        firstName:
          type: string
          description: User first name
          example: "John"
          minLength: 2
          maxLength: 50
        lastName:
          type: string
          description: User last name
          example: "Doe"
          minLength: 2
          maxLength: 50
        role:
          type: string
          enum: [USER, MANAGER, ADMIN]
          description: User role
          example: "USER"
          default: "USER"
      required:
        - username
        - email
        - password
        - firstName
        - lastName
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        type:
          type: string
          description: Token type
          example: "Bearer"
        username:
          type: string
          description: Authenticated username
          example: "admin"
        email:
          type: string
          description: User email
          example: "admin@company.com"
      required:
        - token
        - type
        - username

    # Employee Schemas
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User Authentication
      description: |
        Authenticate user credentials and generate JWT access token for API authorization.
        
        **Test Credentials:**
        - Username: `admin`
        - Password: `admin123`
      operationId: login
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standard_login:
                summary: Standard Admin Login
                value:
                  username: "admin"
                  password: "admin123"
      responses:
        '200':
          description: Authentication successful - JWT token generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed - Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request - Invalid input format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Register a new user account with username, email, password, and role assignment
      operationId: register
      security: []
      requestBody:
        description: User registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              new_user:
                summary: New User Registration
                value:
                  username: "john.doe"
                  email: "john.doe@company.com"
                  password: "SecurePass123!"
                  firstName: "John"
                  lastName: "Doe"
                  role: "USER"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Conflict - Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request - Invalid input or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get User Profile
      description: Retrieve authenticated user's profile information including username, authorities, and authentication status
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                user_profile:
                  summary: User Profile Information
                  value:
                    success: true
                    message: "Profile retrieved successfully"
                    data:
                      username: "john.doe"
                      authorities: ["ROLE_USER"]
                      authenticated: true
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...